requirements.txt
numpy==1.21.4
PyYAML==6.0
torch==1.10.0
tqdm==4.62.3